+++ hip_prof_gen.py
@@ -301,8 +301,6 @@ def generate_prof_header(f, api_map, opts_map):
   f.write('// automatically generated sources\n')
   f.write('#ifndef _HIP_PROF_STR_H\n');
   f.write('#define _HIP_PROF_STR_H\n');
-  f.write('#include <sstream>\n');
-  f.write('#include <string>\n');
   
   # Generating dummy macro for non-public API
   f.write('\n// Dummy API primitives\n')
@@ -383,27 +381,30 @@ def generate_prof_header(f, api_map, opts_map):
   f.write('#define INIT_CB_ARGS_DATA(cb_id, cb_data) INIT_##cb_id##_CB_ARGS_DATA(cb_data)\n')
   
   # Generating the method for the API string, name and parameters
-  f.write('\n')
-  f.write('#if 0\n')
-  f.write('// HIP API string method, method name and parameters\n')
-  f.write('const char* hipApiString(hip_api_id_t id, const hip_api_data_t* data) {\n')
-  f.write('  std::ostringstream oss;\n')
-  f.write('  switch (id) {\n')
-  for name, args in api_map.items():
-    f.write('    case HIP_API_ID_' + name + ':\n')
-    f.write('      oss << "' + name + '("')
-    for ind in range(0, len(args)):
-      arg_tuple = args[ind]
-      arg_name = arg_tuple[1]
-      if ind != 0: f.write(' << ","')
-      f.write('\n          << " ' + arg_name  + '=" << data->args.' + name + '.' + arg_name)
-    f.write('\n          << ")";\n')
-    f.write('    break;\n')
-  f.write('    default: oss << "unknown";\n')
-  f.write('  };\n')
-  f.write('  return strdup(oss.str().c_str());\n')
-  f.write('};\n')
-  f.write('#endif\n')
+  if False:
+    f.write('\n')
+    f.write('#if 0\n')
+    f.write('#include <sstream>\n');
+    f.write('#include <string>\n');
+    f.write('// HIP API string method, method name and parameters\n')
+    f.write('const char* hipApiString(hip_api_id_t id, const hip_api_data_t* data) {\n')
+    f.write('  std::ostringstream oss;\n')
+    f.write('  switch (id) {\n')
+    for name, args in api_map.items():
+      f.write('    case HIP_API_ID_' + name + ':\n')
+      f.write('      oss << "' + name + '("')
+      for ind in range(0, len(args)):
+        arg_tuple = args[ind]
+        arg_name = arg_tuple[1]
+        if ind != 0: f.write(' << ","')
+        f.write('\n          << " ' + arg_name  + '=" << data->args.' + name + '.' + arg_name)
+      f.write('\n          << ")";\n')
+      f.write('    break;\n')
+    f.write('    default: oss << "unknown";\n')
+    f.write('  };\n')
+    f.write('  return strdup(oss.str().c_str());\n')
+    f.write('};\n')
+    f.write('#endif\n')
   
   f.write('#endif  // _HIP_PROF_STR_H\n');
